练习1/requirements.txt
transformers>=4.44.0
torch>=2.2.0
Pillow>=10.0.0
